+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Fixed
+
+* Invalid results for pagination queries when filtering on
+  a part of a composite index
+
 ### Added
 
 * Added jsonpath indexes support for queries